extract searchbar in separate component
