+++ src/components/Landing.tsx
@@ -1,63 +1,17 @@
-import React, { FormEvent, useCallback, useState } from 'react'
-import { CID } from 'multiformats/cid'
 import {
   AppTile,
-  Box,
   LandingPageWrapper,
   LandingPageAppTile,
   LandingPageContent,
-  Label,
   Page,
-  useNetworkName,
-  IconSearch,
-  InputLabelBase,
-  Button,
-  Input,
-  SmartLink
+  SmartLink,
+  useNetworkName
 } from '@glif/react-components'
-import { validateAddressString } from '@glif/filecoin-address'
-import { useRouter } from 'next/router'
-
-import { navigate } from '../utils/urlParams'
-import { PAGE } from '../../constants'
+import SearchBar from './SearchBar'
 
 export default function Landing() {
-  const router = useRouter()
-  const [searchTerm, setSearchTerm] = useState('')
-  const [searchError, setSearchError] = useState('')
-
   const { networkName } = useNetworkName(
     process.env.NEXT_PUBLIC_LOTUS_NODE_JSONRPC
-  )
-
-  const onSubmit = useCallback(
-    (e: FormEvent) => {
-      e.preventDefault()
-      setSearchError('')
-
-      const search = searchTerm.trim()
-
-      try {
-        if (validateAddressString(search)) {
-          navigate(router, {
-            pageUrl: PAGE.ACTOR,
-            newQueryParams: { address: search }
-          })
-          return
-        } else if (CID.parse(searchTerm)) {
-          navigate(router, {
-            pageUrl: PAGE.MESSAGE,
-            newQueryParams: {
-              cid: search
-            }
-          })
-          return
-        }
-      } catch {
-        setSearchError('Invalid search')
-      }
-    },
-    [setSearchError, router, searchTerm]
   )
 
   return (
@@ -84,46 +38,7 @@ export default function Landing() {
         </LandingPageAppTile>
         <LandingPageContent>
           <h2>Search for an address or a transaction hash</h2>
-
-          <form onSubmit={onSubmit}>
-            <Box
-              display='flex'
-              alignItems='center'
-              position='relative'
-              overflow='hidden'
-              maxWidth={13}
-              boxShadow={2}
-              borderRadius={2}
-            >
-              <InputLabelBase display='none' htmlFor='check-fil-address' />
-              <Input.Base
-                id='check-fil-address'
-                pr={8}
-                pl={4}
-                height={7}
-                value={searchTerm}
-                onChange={(e) => setSearchTerm(e.target.value)}
-              />
-              {!searchTerm && <IconSearch position='absolute' left={4} />}
-              <Button
-                position='absolute'
-                right={2}
-                px={4}
-                bg='transparent'
-                type='submit'
-                title='Search'
-                variant='secondary'
-                disabled={!searchTerm}
-              />
-            </Box>
-          </form>
-
-          {searchError && (
-            <Label color='status.fail.background' ml={2} mt={4}>
-              {searchError}
-            </Label>
-          )}
-
+          <SearchBar />
           <p>
             Want to load this app directly from IPFS or Filecoin?
             <br />
@@ -132,7 +47,6 @@ export default function Landing() {
               release page
             </SmartLink>
           </p>
-          
           <p>
             Need help?
             <br />
